% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulate component-wise censored data from an illness-death model}
\usage{
simdat(
  n,
  scale12 = 1/8e-04,
  scale13 = 1/2e-04,
  scale23 = 1/0.0016,
  shape12 = 1,
  shape13 = 1,
  shape23 = 1,
  scale21 = NULL,
  shape21 = 1,
  vital.lfu = c(30.4 * 36, 30.4 * 48),
  visit.schedule = 30.4 * c(6, 12, 18, 24, 30, 36, 42, 48),
  scatter.sd = 10,
  missing.rate = 0,
  visit.rate = NULL,
  renew.param = NULL,
  visit.postprog = 1,
  dependent.visit = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{sample size}

\item{scale12}{The model has states 1, 2 and 3 representing illness free,
alive with illness, and death. \code{scale12} is the scale parameter for the Weibull
distribution used to generate the time from entry in state 1 until transition to state 2.
Default is 1/0.0008.}

\item{scale13}{the scale parameter for the Weibull
distribution used to generate the time from entry in state 1 until transition to state 3.
Default is 1/0.0002.}

\item{scale23}{the scale parameter for the Weibull
distribution used to generate the time from entry in state 2 until transition to state 3.
Default is 1/0.0016.}

\item{shape12}{the shape parameter for the Weibull
distribution used to generate the time from entry in state 1 until transition to state 2.
Default is 1.}

\item{shape13}{the shape parameter for the Weibull
distribution used to generate the time from entry in state 1 until transition to state 3.
Default is 1.}

\item{shape23}{the shape parameter for the Weibull
distribution used to generate the time from entry in state 2 until transition to state 3.
Default is 1.}

\item{scale21}{the scale parameter for the Weibull
distribution used to generate the time from entry in state 2 until transition to state 1.
Default is \code{NULL}, which implies an irreversible illness-death model. If \code{scale21} is
a positive number, the model is reversible, i.e. transitions from state 2 back
to state 1 are allowed. This version of \code{simdat} can only handle reversible models
where all scale parameters are equal to 1.}

\item{shape21}{the shape parameter for the Weibull
distribution used to generate the time from entry in state 2 until transition to state 1.
Default is 1.}

\item{vital.lfu}{defines the right censoring time or distribution. If \code{vital.lfu}
is a vector, right censoring will occur according to a uniform distribution on
the interval defined by the vector. If \code{vital.lfu} is a scalar, all observations
will be right-censored at \code{vital.lfu}.}

\item{visit.schedule}{exactly one of \code{visit.schedule}, \code{visit.rate} and \code{renew.param}
should be non-null If \code{visit.schedule} is non-null, visits will be generated
according to a truncated normal distribution around each time in \code{visit.schedule},
truncated at 3 standard deviations.}

\item{scatter.sd}{the standard deviation of the truncated normal distribution used to
generate visit times}

\item{missing.rate}{a vector of the same length as \code{visit.schedule},
which specifies the proportion of  individuals with missing data at each time
in \code{visit.schedule}. If \code{missing.rate=0} (the default), no visits are missed.}

\item{visit.rate}{exactly one of \code{visit.schedule}, \code{visit.rate} and \code{renew.param}
should be non-null If \code{visit.rate} is non-null,  visits will be generated
according to a Poisson process with rate function \code{visit.rate} as a function of time,
for example, \code{visit.rate = function(t) .0001*t}. For a constant
visit rate, use \code{visit.rate = function(x) sapply(x, function(t) C)} where \code{C}
is the constant rate.}

\item{renew.param}{exactly one of \code{visit.schedule}, \code{visit.rate} and \code{renew.param}
should be non-null If \code{renew.param} is non-null,  visits will be generated
according to a Weibull renewal visit process with shape and scale parameters equal to
the first and second entries of \code{renew.param}.}

\item{visit.postprog}{if equal to 0, visits stop once illness is observed at a visit.
If equal to 1, visits continue as normal. The default is 1.}

\item{dependent.visit}{If \code{dependent.visit} is non-null, an extra visit is
added following an individual's first transition to state 2. If \code{dependent.visit} is
a vector, the time from the transition to state 2 until the visit is generated
from a truncated normal distribution with mean, standard deviation given by the
first and second elements of the vector, respectively.
If \code{dependent.visit} is a function, the  time from the transition to state
2 until the visit will be generated according to a Poisson process with rate
function \code{visit.rate} as a function of time.}

\item{seed}{If \code{seed} is specified, \code{set.seed(seed)} will be run before
generating the data, so the data can be reproduced.}
}
\value{
A dataframe with one row per individual with the variables \code{dtime},
\code{dstatus}, \code{state2obs}, \code{laststate1}, \code{t1}-\code{tm}, \code{x1}-
\code{xm}, and \code{nvisits}.
\code{nvisits} and \code{m} are the largest number of visits observed in the dataset.
\code{dstatus=1} represents death and \code{dstatus=0} represents right censoring.
\code{ti} and \code{xi} are the time and observed status at the \code{i}-th visit.
\code{xi=1} means alive and event free (state 1 in a multistate illness-death model),
and \code{xi=2} means alive with event. \code{state2obs} is the time of the first visit
in state 2 (\code{Inf} if state 2 was never observed) and \code{laststate1} is the time
of the last observation in state 1 before state 2 was observed (\code{0} if an
individual was already in state 2 at the first visit). Visits are numbered by
 ascending time and missing visits do not get a placeholder. That is, if
 \code{visit.schedule} was 30, 60, 90 and 120 days, but the actual visit times
 for an individual were \code{45,40, NA, 125} because of normal scatter and missing visits,
 the entries \code{t1-t4} will be \code{40, 45, 125, NA}.
}
\description{
This function simulates component-wise censored data from an irreversible or
reversible illness-death model. The data is component-wise censored because
illness status is intermittently observed at visits, while vital status (alive/dead)
is fully observed until right censoring time. All individuals start alive and illness-free.
Time from state entry to transition is simulated from a Weibull distribution with
user-specified parameters, and the user specifies the visit process.
}
\examples{
simdat(50, scale12=1/.0008, scale13=1/.0002, scale23=1/.0016,
vital.lfu=c(30.4*36, 30.4*48),
visit.schedule = 30.4*c(6, 12, 18, 24, 30, 36, 42, 48),
scatter.sd=10)
}
